pyproject.toml
[tool.poetry]
name = "testtaskma"
version = "0.1.0"
description = ""
authors = ["Semyon Albeev <semyon.albeev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.4.0"
python-multipart = "^0.0.9"
aioboto3 = "^13.1.1"
pillow = "^10.4.0"
python-magic = "^0.4.27"
libmagic = "^1.0"
python-magic-bin = "^0.4.14"
mutagen = "^1.47.0"
pypdf2 = "^3.0.1"
python-docx = "^1.1.2"
moviepy = "^1.0.3"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
